add conftest script
