+++ .pre-commit-config.yaml
@@ -17,3 +17,7 @@ repos:
     -   id: detect-private-key
     -   id: name-tests-test
     -   id: requirements-txt-fixer
+-   repo: https://github.com/asottile/reorder_python_imports
+    rev: v3.2.0
+    hooks:
+    -   id: reorder-python-imports
